+++ radix-engine/src/blueprints/package/mod.rs
@@ -1,7 +1,5 @@
 mod package;
-mod package_extractor;
 mod substates;
 
 pub use package::*;
-pub use package_extractor::*;
 pub use substates::*;
