+++ radix-engine/src/lib.rs
@@ -28,3 +28,5 @@ pub mod wasm;
 pub mod types;
 
 pub mod errors;
+
+pub mod utils;
