+++ radix-engine/src/utils/mod.rs
@@ -1,0 +1,3 @@
+mod package_extractor;
+
+pub use package_extractor::*;
