+++ radix-engine/src/utils/package_extractor.rs
