+++ simulator/src/resim/error.rs
@@ -2,9 +2,9 @@ use std::io;
 use std::path::PathBuf;
 
 use radix_engine::errors::{RejectionError, RuntimeError};
-use radix_engine::system::package::ExtractSchemaError;
 use radix_engine::transaction::AbortReason;
 use radix_engine::types::{AddressError, ComponentAddress, PackageAddress};
+use radix_engine::utils::ExtractSchemaError;
 use radix_engine::wasm::PrepareError;
 use radix_engine_interface::blueprints::resource::ParseNonFungibleGlobalIdError;
 use radix_engine_interface::network::ParseNetworkError;
